chore: split `types` and `utils`
